+++ modules/core/src/main/java/org/apache/ignite/mxbean/TransactionsMXBean.java
@@ -152,7 +152,7 @@ public interface TransactionsMXBean {
         "Returns the percentage of samples of long running transactions that will be dumped in log, if " +
         "longTransactionTimeDumpThreshold is set to non-zero value."
     )
-    public float getLongTransactionTimeDumpSampleLimit();
+    public double getLongTransactionTimeDumpSampleLimit();
 
     @MXBeanDescription(
         "Sets the percentage of samples of long running transactions that will be dumped in log, if " +
@@ -160,5 +160,5 @@ public interface TransactionsMXBean {
     )
     @MXBeanParametersNames("percentage")
     @MXBeanParametersDescriptions("Samples percentage.")
-    public void setLongTransactionTimeDumpSampleLimit(float percentage);
+    public void setLongTransactionTimeDumpSampleLimit(double percentage);
 }
